feature/050426 refactor ventas

+++ src/main/java/com/svc/ventas/models/mapstruct/dto/ProductoDetalleVentaDto.java
@@ -17,7 +17,7 @@ public class ProductoDetalleVentaDto {
 
 	private String nombre;
 
-	private Integer cantidad;
+	private BigDecimal cantidad;
 
 	private BigDecimal precioVenta;
